Compiler warning
(cherry picked from commit 7bd54a700a39c6a15fb222d5cf696f26ab750e22)

Source commit: 1b39d0907ee953ecf786bb2b4048641b11abc19b

+++ libs/libs3/src/request.c
@@ -1314,7 +1314,7 @@ static S3Status setup_neon(Request *request,
         {
             *colon = '\0';
         }
-        for (int i = 0; i < strlen(params->bucketContext.bucketName); i++)
+        for (size_t i = 0; i < strlen(params->bucketContext.bucketName); i++)
         {
             if (hostName[i] == '.')
             {
